Consolidate PackageSettings
