--- src/test/java/io/github/ygojson/application/yugipedia/parser/acceptance/PackageSettings.java
@@ -1,9 +1,0 @@
-package io.github.ygojson.application.yugipedia.parser.acceptance;
-
-/**
- * Settings for the approval tests configuration.
- */
-public class PackageSettings {
-
-	public static String ApprovalBaseDirectory = "../resources";
-}
